backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
urllib3
pandas
numpy
